function_app/requirements.txt
azure-functions==1.21.3
azure-cosmos==4.9.0
openai==1.59.9
